Add files via upload
The Great Refactoring is complete. In particular, all global variables have been replaced by singleton classes. The program is back to working normally. Improved reverse file functionality so if a particular sound modification is already on file, that file is reloaded instead of generating a new file from scratch. The Reversed Label now says "Unreversed file!" if the original file is reloaded (by clicking the Reverse File button an even number of times).